+++ NEWS.md
@@ -1,6 +1,13 @@
 Revision history for the R/broman package
 -----------------------------------------
 
+## Version 0.85-1, 2025-05-31
+
+- Added `runningratio2()` which is like `runningratio()` but instead
+  of a fixed-width window, it uses an adaptive window to hit some
+  target denominator.
+
+
 ## Version 0.84, 2024-05-18
 
 - Allow `NA`s in `runningmean()` and `runningratio()`.